Configure JWT auth middleware through options struct
